# Run optimize command under `filament` key in `artisan optimize`