--- a/packages/support/src/SupportServiceProvider.php
+++ b/packages/support/src/SupportServiceProvider.php
@@ -154,7 +154,7 @@ class SupportServiceProvider extends PackageServiceProvider
                 $this->optimizes(
                     optimize: 'filament:optimize', /** @phpstan-ignore-line */
                     clear: 'filament:optimize-clear', /** @phpstan-ignore-line */
-                    key: 'filament'
+                    key: 'filament', /** @phpstan-ignore-line */
                 );
             }
         }